refactor(frontend): extract NodeCardStackList from NodeCard (F5-9)
